+++ R/portion.R
@@ -22,7 +22,7 @@
 #'
 #' \code{portion_R} and \code{portion_IR} can be used to calculate resistance, \code{portion_S} and \code{portion_SI} can be used to calculate susceptibility.\cr
 #' @param ... one or more vectors (or columns) with antibiotic interpretations. They will be transformed internally with \code{\link{as.rsi}} if needed. Use multiple columns to calculate (the lack of) co-resistance: the probability where one of two drugs have a resistant or susceptible result. See Examples.
-#' @param minimum minimal amount of available isolates. Any number lower than \code{minimum} will return \code{NA}. The default number of \code{30} isolates is advised by the CLSI as best practice, see Source.
+#' @param minimum minimal amount of available isolates. Any number lower than \code{minimum} will return \code{NA} with a warning. The default number of \code{30} isolates is advised by the Clinical and Laboratory Standards Institute (CLSI) as best practice, see Source.
 #' @param as_percent logical to indicate whether the output must be returned as a hundred fold with \% sign (a character). A value of \code{0.123456} will then be returned as \code{"12.3\%"}.
 #' @param data a \code{data.frame} containing columns with class \code{rsi} (see \code{\link{as.rsi}})
 #' @param translate_ab a column name of the \code{\link{antibiotics}} data set to translate the antibiotic abbreviations to, using \code{\link{abname}}. This can be set with \code{\link{getOption}("get_antibiotic_names")}.
@@ -50,8 +50,7 @@
 #' @source \strong{M39 Analysis and Presentation of Cumulative Antimicrobial Susceptibility Test Data, 4th Edition}, 2014, \emph{Clinical and Laboratory Standards Institute (CLSI)}. \url{https://clsi.org/standards/products/microbiology/documents/m39/}.
 #'
 #' Wickham H. \strong{Tidy Data.} The Journal of Statistical Software, vol. 59, 2014. \url{http://vita.had.co.nz/papers/tidy-data.html}
-#' @seealso \code{\link[AMR]{count}_*} to count resistant and susceptibile isolates.\cr
-#' \code{\link{n_rsi}} to count all cases where antimicrobial results are available.
+#' @seealso \code{\link[AMR]{count}_*} to count resistant and susceptibile isolates.
 #' @keywords resistance susceptibility rsi_df rsi antibiotics isolate isolates
 #' @return Double or, when \code{as_percent = TRUE}, a character.
 #' @rdname portion
@@ -92,24 +91,24 @@
 #'
 #' # Calculate co-resistance between amoxicillin/clav acid and gentamicin,
 #' # so we can see that combination therapy does a lot more than mono therapy:
-#' septic_patients %>% portion_S(amcl)       # S = 67.3%
-#' septic_patients %>% n_rsi(amcl)           # n = 1570
+#' septic_patients %>% portion_S(amcl)       # S = 67.1%
+#' septic_patients %>% count_all(amcl)       # n = 1576
 #'
 #' septic_patients %>% portion_S(gent)       # S = 74.0%
-#' septic_patients %>% n_rsi(gent)           # n = 1842
+#' septic_patients %>% count_all(gent)       # n = 1855
 #'
-#' septic_patients %>% portion_S(amcl, gent) # S = 92.1%
-#' septic_patients %>% n_rsi(amcl, gent)     # n = 1504
+#' septic_patients %>% portion_S(amcl, gent) # S = 92.0%
+#' septic_patients %>% count_all(amcl, gent) # n = 1517
 #'
 #'
 #' septic_patients %>%
 #'   group_by(hospital_id) %>%
 #'   summarise(cipro_p = portion_S(cipr, as_percent = TRUE),
-#'             cipro_n = n_rsi(cipr),
+#'             cipro_n = count_all(cipr),
 #'             genta_p = portion_S(gent, as_percent = TRUE),
-#'             genta_n = n_rsi(gent),
+#'             genta_n = count_all(gent),
 #'             combination_p = portion_S(cipr, gent, as_percent = TRUE),
-#'             combination_n = n_rsi(cipr, gent))
+#'             combination_n = count_all(cipr, gent))
 #'
 #' # Get portions S/I/R immediately of all rsi columns
 #' septic_patients %>%
@@ -130,7 +129,7 @@
 #'   filter(first_isolate == TRUE,
 #'          genus == "Helicobacter") %>%
 #'   summarise(p = portion_S(amox, metr),  # amoxicillin with metronidazole
-#'             n = n_rsi(amox, metr))
+#'             n = count_all(amox, metr))
 #' }
 portion_R <- function(...,
                       minimum = 30,
@@ -273,6 +272,8 @@ rsi <- function(ab1,
                 as_percent = FALSE,
                 ...) {
 
+  .Deprecated(new = paste0("portion_", interpretation))
+
   if (all(is.null(ab2))) {
     df <- tibble(ab1 = ab1)
   } else {
@@ -280,19 +281,16 @@ rsi <- function(ab1,
                  ab2 = ab2)
   }
 
+  if (!interpretation %in% c("S", "SI", "IS", "I", "RI", "IR", "R")) {
+    stop("invalid interpretation")
+  }
+
   result <- case_when(
     interpretation == "S"             ~ portion_S(df, minimum = minimum, as_percent = FALSE),
     interpretation %in% c("SI", "IS") ~ portion_SI(df, minimum = minimum, as_percent = FALSE),
     interpretation == "I"             ~ portion_I(df, minimum = minimum, as_percent = FALSE),
     interpretation %in% c("RI", "IR") ~ portion_IR(df, minimum = minimum, as_percent = FALSE),
-    interpretation == "R"             ~ portion_R(df, minimum = minimum, as_percent = FALSE),
-    TRUE ~ -1
-  )
-  if (result == -1) {
-    stop("invalid interpretation")
-  }
-
-  .Deprecated(new = paste0("portion_", interpretation))
+    interpretation == "R"             ~ portion_R(df, minimum = minimum, as_percent = FALSE))
 
   if (as_percent == TRUE) {
     percent(result, force_zero = TRUE)