+++ tests/testthat/test-atc.R
@@ -10,8 +10,8 @@ test_that("atc_property works", {
     expect_equal(atc_property("J01CA04", property = "DDD"),
                  atc_ddd("J01CA04"))
 
-    expect_identical(atc_property("J01CA04", property = "Groups"),
-                     atc_groups("J01CA04"))
+    # expect_identical(atc_property("J01CA04", property = "Groups"),
+    #                  atc_groups("J01CA04"))
 
     expect_warning(atc_property("ABCDEFG", property = "DDD"))